components/indexer/requirements.txt
aiohttp==3.13.3
asyncpg==0.31.0
cryptg==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
qdrant-client==1.17.0
telethon==1.42.0